requirements.txt
django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.0.0
django-oauth-toolkit

channels==2.4.0
channels_redis==2.4.2
daphne==2.4.1
gunicorn==20.0.4

psycopg2-binary==2.8.5
whitenoise==5.0.1
django-heroku==0.3.1

redis==3.4.1
coverage==5.0.3

requests==2.23.0
python-dotenv==0.12.0

pybreaker==0.6.0